requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
pymongo==4.10.1
redis==5.0.8
langchain==0.2.14
langchain-core==0.2.37
langchain-community==0.2.6
langchain-google-genai==1.0.5
google-generativeai==0.5.3
pydantic==2.9.2
requests==2.32.3
tqdm==4.66.5
tenacity==8.5.0
psycopg[binary]
